change batch to log
